Chore: cleaning structure
