+++ src/components/orderBook/index.tsx
@@ -3,8 +3,8 @@ import { Flex, HStack, Table, Tbody, Td, Th, Thead, Tr } from '@chakra-ui/react'
 import { uid } from 'uid';
 import * as React from 'react';
 import { useAppSelector } from '../../hooks/hooks';
-import { Order } from '../order/order';
-import { Depth } from '../../utils';
+import { Order } from '../order';
+import { Depth } from '../../../helpers';
 import { useTranslations } from 'next-intl';
 type Props = {
   type: 'bid' | 'ask';
