+++ src/pages/_app.tsx
@@ -2,7 +2,7 @@ import '@/styles/globals.css'
 import type { AppProps } from 'next/app'
 import { ChakraProvider } from '@chakra-ui/react'
 import { Provider } from 'react-redux'
-import store from '@/state/store'
+import store from '@/redux/store'
 import { NextIntlProvider } from 'next-intl'
 
 export default function App({ Component, pageProps }: AppProps) {
